+++ qcommon/cmodel.c
@@ -1558,7 +1558,7 @@ void CM_DecompressVis (byte *in, byte *out)
 		in += 2;
 		if ((out_p - out) + c > row)
 		{
-			c = row - (out_p - out);
+			c = (int)(row - (out_p - out));
 			Com_DPrintf ("warning: Vis decompression overrun\n");
 		}
 		while (c)
