+++ qcommon/net_chan.c
@@ -142,7 +142,7 @@ void Netchan_OutOfBandPrint (int net_socket, netadr_t adr, char *format, ...)
 #endif /* __APPLE__ || MACOSX */
 	va_end (argptr);
 
-	Netchan_OutOfBand (net_socket, adr, strlen(string), (byte *)string);
+	Netchan_OutOfBand (net_socket, adr, (int)strlen(string), (byte *)string);
 }
